🐛 Fixed the issue of "when getting models in batches, I hope to automatically refresh according to the model type" #1049
